Refactored: is done method tested

+++ src/main/java/biz/interretis/newinjava/futures/ComputationsSimulator.java
@@ -1,15 +1,14 @@
 package biz.interretis.newinjava.futures;
 
-
 public class ComputationsSimulator {
 
-    public static final int ANSWER_TO_ULTIMATE_QUESTION = 42;
+    public static final double ANSWER_TO_ULTIMATE_QUESTION = 42.0;
 
     public static Double doSomeLongComputation(final long millis) {
 
         delay(millis);
 
-        return (double) ANSWER_TO_ULTIMATE_QUESTION;
+        return ANSWER_TO_ULTIMATE_QUESTION;
     }
 
     public static void doSomethingElse(final long millis) {
